requirements.txt
Flask
Flask-PyMongo
Flask-Mail
python-dotenv
